fix(workflows): Update EC2 target path and fix typo
The `ec2-setup.yaml` workflow's target path for deployment was updated to ensure consistent directory handling. A trailing slash was added to `/home/${{ secrets.EC2_USERNAME }}/app` to explicitly denote it as a directory.

Additionally, a typo in the `frontend-cd.yaml` workflow was corrected, changing "SSH inro EC2 and Deploy" to "SSH into EC2 and Deploy" for improved clarity.